+++ account_payment_order/models/account_invoice.py
@@ -52,22 +52,7 @@ class AccountInvoice(models.Model):
         return res
 
     @api.multi
-    def _prepare_new_payment_order(self, payment_mode=None):
-        self.ensure_one()
-        if payment_mode is None:
-            payment_mode = self.env['account.payment.mode']
-        vals = {
-            'payment_mode_id': payment_mode.id or self.payment_mode_id.id,
-        }
-        # other important fields are set by the inherit of create
-        # in account_payment_order.py
-        return vals
-
-    @api.multi
     def create_account_payment_line(self):
-        apoo = self.env['account.payment.order']
-        result_payorder_ids = []
-        action_payment_type = 'debit'
         for inv in self:
             if inv.state != 'open':
                 raise UserError(_(
@@ -75,65 +60,8 @@ class AccountInvoice(models.Model):
             if not inv.move_id:
                 raise UserError(_(
                     "No Journal Entry on invoice %s") % inv.number)
-            applicable_lines = inv.move_id.line_ids.filtered(
-                lambda x: (
-                    not x.reconciled and x.payment_mode_id.payment_order_ok and
-                    x.account_id.internal_type in ('receivable', 'payable') and
-                    not any(p_state in ('draft', 'open', 'generated')
-                            for p_state in x.payment_line_ids.mapped('state'))
-                )
-            )
-            if not applicable_lines:
-                raise UserError(_(
-                    'No Payment Line created for invoice %s because '
-                    'it already exists or because this invoice is '
-                    'already paid.') % inv.number)
-            payment_modes = applicable_lines.mapped('payment_mode_id')
-            if not payment_modes:
-                raise UserError(_(
-                    "No Payment Mode on invoice %s") % inv.number)
-            for payment_mode in payment_modes:
-                payorder = apoo.search([
-                    ('payment_mode_id', '=', payment_mode.id),
-                    ('state', '=', 'draft')
-                ], limit=1)
-                new_payorder = False
-                if not payorder:
-                    payorder = apoo.create(inv._prepare_new_payment_order(
-                        payment_mode
-                    ))
-                    new_payorder = True
-                result_payorder_ids.append(payorder.id)
-                action_payment_type = payorder.payment_type
-                count = 0
-                for line in applicable_lines.filtered(
-                    lambda x: x.payment_mode_id == payment_mode
-                ):
-                    line.create_payment_line_from_move_line(payorder)
-                    count += 1
-                if new_payorder:
-                    inv.message_post(body=_(
-                        '%d payment lines added to the new draft payment '
-                        'order %s which has been automatically created.')
-                        % (count, payorder.name))
-                else:
-                    inv.message_post(body=_(
-                        '%d payment lines added to the existing draft '
-                        'payment order %s.')
-                        % (count, payorder.name))
-        action = self.env['ir.actions.act_window'].for_xml_id(
-            'account_payment_order',
-            'account_payment_order_%s_action' % action_payment_type)
-        if len(result_payorder_ids) == 1:
-            action.update({
-                'view_mode': 'form,tree,pivot,graph',
-                'res_id': payorder.id,
-                'views': False,
-                })
-        else:
-            action.update({
-                'view_mode': 'tree,form,pivot,graph',
-                'domain': "[('id', 'in', %s)]" % result_payorder_ids,
-                'views': False,
-                })
-        return action
+
+        return self.mapped("move_id.line_ids").filtered(
+            lambda x: x.account_id.internal_type in (
+                'receivable', 'payable')).action_add_to_payment_order()
+
